Add Garden condition alert and condition & last operation dashboard
d7b9f90 introduced new metrics tracking the last operation and conditions.
This commit introduces a new alert if the Garden resource reports
conditions that are not `True` for 10 minutes.
It also adds a supplemental dashboard that displays the state of the Garden
resource's conditions and the last operation over time.